requirements.txt
flask==3.0.2
pandas==2.0.1
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
flask-cors==4.0.0
gunicorn==21.2.0